Properly adapt to innacurate guns
Look at new enemies that damage us.

Start camping ( crouching + shooting + not moving ) in more parts of the 'brain' when we happen to get a good weapon

Don't run our path if we miss alot of shots.
And start crouching if we miss alot of shots with a really good gun.

Overcharged bot is now the only one with perfect weapon proficiency ( still misses a load of shots anyways )

Some error stomps